+++ test_sirikit_app/MessageIntentUI/Info.plist
@@ -9,6 +9,7 @@
 			<key>IntentsSupported</key>
 			<array>
 				<string>INSendMessageIntent</string>
+				<string>SearchToolIntent</string>
 			</array>
 		</dict>
 		<key>NSExtensionMainStoryboard</key>
