+++ test_sirikit_app/test_sirikit_app/AppDelegate.swift
@@ -8,6 +8,7 @@
 import Foundation
 import UIKit
 
+// https://medium.com/%E5%BD%BC%E5%BE%97%E6%BD%98%E7%9A%84-swift-ios-app-%E9%96%8B%E7%99%BC%E5%95%8F%E9%A1%8C%E8%A7%A3%E7%AD%94%E9%9B%86/swiftui-%E5%A6%82%E4%BD%95%E6%90%AD%E9%85%8D-appdelegate-%E5%88%A9%E7%94%A8-uiapplicationdelegateadaptor-96446be3e17
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
         print("your code here")
